# influencer-platform-api/src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Your API",
    "description": "Generated by swagger-autogen",
    "version": "1.0.0"
  },
  "host": "localhost:5000/api",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/campaign/": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}